Correccion CRUDS, y CRUD PODER, METAHUMANO Y METAPODER, TERMINADOS
