+++ package.json
@@ -11,16 +11,18 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@mikro-orm/cli": "5.8.10",
     "@mikro-orm/core": "5.8.10",
+    "@mikro-orm/migrations": "5.8.10",
     "@mikro-orm/mysql": "5.8.10",
     "@mikro-orm/sql-highlighter": "1.0.1",
-    "express": "4.18.2",
+    "express": "^4.18.2",
     "reflect-metadata": "0.1.13"
   },
   "devDependencies": {
-    "@types/express": "4.17.17",
+    "@types/express": "^4.17.17",
     "@types/node": "20.5.0",
     "tsc-watch": "6.0.4",
-    "typescript": "5.1.6"
+    "typescript": "^5.1.6"
   }
 }
