+++ src/app.ts
@@ -4,15 +4,13 @@ import { RequestContext } from '@mikro-orm/core'
 import { orm, syncSchema } from './shared/db/orm.js'
 
 import metahumanosRoutes from './metahumano/metahumano.routes.js'
-import poderesRoutes from './metahumano/poder.routes.js'
+import poderesRoutes from './poder/poder.routes.js'
+import metaPoderesRoutes from './metaPoder/metaPoder.routes.js'
 import { burocratasRouter } from './Burocratas/Burocrata.routes.js'
 import { multasRouter } from './Multas/Multa.routes.js'
 
 const app = express()
 app.use(express.json())
-
-
-
 
 
 // Contexto de EntityManager por request
@@ -24,6 +22,7 @@ app.use((req, res, next) => {
 // Rutas
 app.use('/api/metahumanos', metahumanosRoutes)
 app.use('/api/poderes', poderesRoutes)
+app.use('/api/metapoderes', metaPoderesRoutes)
 app.use('/api/Burocratas', burocratasRouter)
 app.use('/api/Multas', multasRouter)
