+++ src/poder/poder.entity.ts
@@ -5,7 +5,8 @@ import {
   Collection,
 } from '@mikro-orm/core'
 import { BaseEntity } from '../shared/db/baseEntity.entity.js'
-import { MetaPoder } from './metaPoder.entity.js'
+import { MetaPoder } from '../metaPoder/metaPoder.entity.js'
+
 
 @Entity()
 export class Poder extends BaseEntity {
@@ -30,3 +31,4 @@ export class Poder extends BaseEntity {
   @OneToMany(() => MetaPoder, (metaPoder) => metaPoder.poder)
   metahumanos = new Collection<MetaPoder>(this)
 }
+
